fix python tests + added python class SimpleFoxgloveServerListener()
