+++ python/src/foxglove_websocket/examples/threaded_server/__main__.py
@@ -3,8 +3,8 @@ import logging
 import concurrent.futures
 from typing import Any, Coroutine, Dict, Optional
 from foxglove_websocket import run_cancellable
-from foxglove_websocket.server import FoxgloveServer, FoxgloveServerListener
-from foxglove_websocket.types import ClientChannel, ChannelId, ChannelWithoutId
+from foxglove_websocket.server import FoxgloveServer, SimpleFoxgloveServerListener
+from foxglove_websocket.types import ChannelId, ChannelWithoutId
 
 from .middleware import ExampleMiddlewareThread, MiddlewareChannelId
 
@@ -118,7 +118,7 @@ async def main():
 
         # Configure callbacks that the FoxgloveServer will call when its clients subscribe or
         # unsubscribe from our channels.
-        class Listener(FoxgloveServerListener):
+        class Listener(SimpleFoxgloveServerListener):
             def on_subscribe(self, server: FoxgloveServer, channel_id: ChannelId):
                 """
                 When a WebSocket client first subscribes to a channel, create a subscription in the middleware.
@@ -132,22 +132,6 @@ async def main():
                 """
                 logger.info("Last client unsubscribed from %d", channel_id)
                 middleware.handle_unsubscribe_threadsafe(reverse_id_map[channel_id])
-
-            def on_client_advertise(
-                self, server: "FoxgloveServer", channel: ClientChannel
-            ):
-                pass
-
-            def on_client_unadvertise(self, server: "FoxgloveServer", topic: str):
-                pass
-
-            def on_client_data(
-                self,
-                server: "FoxgloveServer",
-                data: Dict[str, Any],
-                timestamp: Optional[int] = None,
-            ):
-                pass
 
         server.set_listener(Listener())
